video photo and db code
just added the function that will enter or edit a peson's record in the db.
also added the media folder where we're going to be keeping the video and photo files for the person's profile.
I also added the video and phot columns to the db which i had forgotten about before for some reason